+++ lib/deep/set.ts
@@ -1,7 +1,48 @@
 'use strict';
 
-//  Set a value for a path in a json-like structure
-export default function deepSet (obj, path, value = null, define = false) {
+/**
+ * Sets a property and its value deep in the structure of an object
+ *
+ * Example:
+ *  const myObj = {a: 2};
+ *  deepSet(myObj, 'b.c.d.e', 4);
+ * Output:
+ *  {
+ *      a: 2,
+ *      b: {
+ *          c: {
+ *              d: {
+ *                  e: 4
+ *              }
+ *          }
+ *      }
+ *  }
+ *
+ * Example:
+ *  const myObj = {a: 2, b: [{price: 2}]};
+ *  deepSet(myObj, 'b[0].price', 4);
+ * Output:
+ *  {
+ *      a: 2,
+ *      b: [
+ *          {price: 4}
+ *      ]
+ *  }
+ *
+ * @param val - Object to set the value on
+ * @param path - Path string to deeply set the value at
+ * @param value - Value to set (if using defineProperty can be an accessor object)
+ * @param define - Whether or not the property should be directly assigned or set using Object.defineProperty
+ *
+ * @returns True or false depending on whether or not the property was set correctly
+ * @throws {TypeError}
+ */
+export default function deepSet (
+    obj:{[key:string]:any},
+    path:string,
+    value:any,
+    define:boolean=false
+) {
     if (
         Object.prototype.toString.call(obj) !== '[object Object]' &&
         !Array.isArray(obj)
@@ -11,7 +52,7 @@ export default function deepSet (obj, path, value = null, define = false) {
     if (typeof path !== 'string') throw new TypeError('No path was given');
 
     //  Check if path contains rejected keys
-    if (/__proto__|constructor|prototype/.test(path)) throw new Error('Malicious path provided');
+    if (/__proto__|constructor|prototype/.test(path)) throw new TypeError('Malicious path provided');
 
     //  Check if path contains content
     const path_s = path.trim();